Fix : reduce complexity of ChatFeedRecyclerAdapter
Fix : reduce complexity of chatfeedrecycleradapter

Fix : reduce complexity of chatfeedrecycleradapter
